+++ packages/ical-tools/src/lib/vevent.ts
@@ -1,14 +1,19 @@
 import {
   add,
   ArgumentError,
+  CollectionIterator,
   DateAdapter,
   DateInput,
   Dates,
   DateTime,
+  ICollectionsArgs,
+  Include,
+  IOccurrencesArgs,
   IProvidedRuleOptions,
   IRunArgs,
   IScheduleLike,
   OccurrenceGenerator,
+  OccurrenceIterator,
   OccurrenceStream,
   Omit,
   OperatorFnOutput,
@@ -20,26 +25,35 @@ import {
 
 const VEVENT_ID = Symbol.for('b1666600-db88-4d8e-9e40-05fdbc48d650');
 
+type GetVEventType<T> = Include<T, VEvent<any, any, any, any>> extends never
+  ? VEvent<any, any, any, any>
+  : Include<T, VEvent<any, any, any, any>>;
+
 export type IVEventRuleOptions<T extends typeof DateAdapter> = Omit<
   Omit<IProvidedRuleOptions<T>, 'start'>,
   'duration'
 >;
 
-export class VEvent<T extends typeof DateAdapter, D = any> extends OccurrenceGenerator<T>
-  implements IScheduleLike<T> {
+export class VEvent<
+  T extends typeof DateAdapter,
+  D = any,
+  RR extends Rule<T> = Rule<T>,
+  RD extends Dates<T> = Dates<T>
+> extends OccurrenceGenerator<T> implements IScheduleLike<T> {
   /**
    * Similar to `Array.isArray`, `isVEvent` provides a surefire method
    * of determining if an object is a `VEvent` by checking against the
    * global symbol registry.
    */
-  static isVEvent(object: any): object is VEvent<any> {
+  // @ts-ignore the check is working as intended but typescript doesn't like it for some reason
+  static isVEvent<T>(object: T): object is GetVEventType<T> {
     return !!(object && typeof object === 'object' && (object as any)[VEVENT_ID]);
   }
 
   // For some reason, error is thrown if typed as `readonly Rule<T>[]`
-  readonly rrules: ReadonlyArray<Rule<T>> = [];
+  readonly rrules: ReadonlyArray<RR> = [];
   readonly exrules: ReadonlyArray<Rule<T>> = [];
-  readonly rdates: Dates<T>;
+  readonly rdates: RD;
   readonly exdates: Dates<T>;
 
   pipe: (...operatorFns: OperatorFnOutput<T>[]) => OccurrenceStream<T> = pipeFn(this);
@@ -62,9 +76,9 @@ export class VEvent<T extends typeof DateAdapter, D = any> extends OccurrenceGen
     start: DateInput<T>;
     dateAdapter?: T;
     data?: D;
-    rrules?: ReadonlyArray<IVEventRuleOptions<T> | Rule<T>>;
+    rrules?: ReadonlyArray<IVEventRuleOptions<T> | RR>;
     exrules?: ReadonlyArray<IVEventRuleOptions<T> | Rule<T>>;
-    rdates?: ReadonlyArray<DateInput<T>> | Dates<T>;
+    rdates?: ReadonlyArray<DateInput<T>> | RD;
     exdates?: ReadonlyArray<DateInput<T>> | Dates<T>;
   }) {
     super(args);
@@ -81,7 +95,7 @@ export class VEvent<T extends typeof DateAdapter, D = any> extends OccurrenceGen
     if (args.rrules) {
       this.rrules = args.rrules.map(ruleArgs => {
         if (Rule.isRule(ruleArgs)) {
-          if (!this.normalizeDateInput(ruleArgs.options.start).isEqual(this._start)) {
+          if (!this.normalizeDateInput((ruleArgs as RR).options.start).isEqual(this._start)) {
             throw new ArgumentError(
               'When passing a `Rule` object to the `VEvent` constructor, ' +
                 'the rule `start` time must equal the `VEvent` start time.',
@@ -93,12 +107,12 @@ export class VEvent<T extends typeof DateAdapter, D = any> extends OccurrenceGen
             );
           }
 
-          return ruleArgs;
+          return ruleArgs as RR;
         } else {
           return new Rule(standardizeVEventRuleOptions(ruleArgs as IVEventRuleOptions<T>, args), {
             dateAdapter: this.dateAdapter as any,
             timezone: this.timezone,
-          });
+          }) as RR;
         }
       });
     }
@@ -130,14 +144,14 @@ export class VEvent<T extends typeof DateAdapter, D = any> extends OccurrenceGen
 
     if (args.rdates) {
       this.rdates = Dates.isDates(args.rdates)
-        ? args.rdates.set('timezone', this.timezone)
-        : new Dates({
+        ? (args.rdates.set('timezone', this.timezone) as RD)
+        : (new Dates({
             dates: args.rdates as ReadonlyArray<DateInput<T>>,
             dateAdapter: this.dateAdapter,
             timezone: this.timezone,
-          });
+          }) as RD);
     } else {
-      this.rdates = new Dates({ dateAdapter: this.dateAdapter, timezone: this.timezone });
+      this.rdates = new Dates({ dateAdapter: this.dateAdapter, timezone: this.timezone }) as RD;
     }
 
     if (args.exdates) {
@@ -177,8 +191,16 @@ export class VEvent<T extends typeof DateAdapter, D = any> extends OccurrenceGen
     });
   }
 
-  add(prop: 'rrule' | 'exrule', value: Rule<T, unknown>): VEvent<T, D>;
-  add(prop: 'rdate' | 'exdate', value: DateInput<T>): VEvent<T, D>;
+  occurrences(args: IOccurrencesArgs<T> = {}): OccurrenceIterator<T, [this, RR | RD]> {
+    return new OccurrenceIterator(this, this.normalizeOccurrencesArgs(args));
+  }
+
+  collections(args: ICollectionsArgs<T> = {}): CollectionIterator<T, [this, RR | RD]> {
+    return new CollectionIterator(this, this.normalizeCollectionsArgs(args));
+  }
+
+  add(prop: 'rrule' | 'exrule', value: Rule<T, unknown>): VEvent<T, D, RR, RD>;
+  add(prop: 'rdate' | 'exdate', value: DateInput<T>): VEvent<T, D, RR, RD>;
   add(prop: 'rdate' | 'exdate' | 'rrule' | 'exrule', value: Rule<T, unknown> | DateInput<T>) {
     const rrules = this.rrules.slice();
     const exrules = this.exrules.slice();
@@ -187,13 +209,13 @@ export class VEvent<T extends typeof DateAdapter, D = any> extends OccurrenceGen
 
     switch (prop) {
       case 'rrule':
-        rrules.push(value as Rule<T, unknown>);
+        rrules.push(value as RR);
         break;
       case 'exrule':
         exrules.push(value as Rule<T, unknown>);
         break;
       case 'rdate':
-        rdates = this.rdates.add(value as DateInput<T>);
+        rdates = this.rdates.add(value as DateInput<T>) as RD;
         break;
       case 'exdate':
         exdates = this.exdates.add(value as DateInput<T>);
@@ -211,8 +233,8 @@ export class VEvent<T extends typeof DateAdapter, D = any> extends OccurrenceGen
     });
   }
 
-  remove(prop: 'rrule' | 'exrule', value: Rule<T, unknown>): VEvent<T, D>;
-  remove(prop: 'rdate' | 'exdate', value: DateInput<T>): VEvent<T, D>;
+  remove(prop: 'rrule' | 'exrule', value: Rule<T, unknown>): VEvent<T, D, RR, RD>;
+  remove(prop: 'rdate' | 'exdate', value: DateInput<T>): VEvent<T, D, RR, RD>;
   remove(prop: 'rdate' | 'exdate' | 'rrule' | 'exrule', value: Rule<T, unknown> | DateInput<T>) {
     let rrules = this.rrules;
     let exrules = this.exrules;
@@ -227,7 +249,7 @@ export class VEvent<T extends typeof DateAdapter, D = any> extends OccurrenceGen
         exrules = exrules.filter(rule => rule !== value);
         break;
       case 'rdate':
-        rdates = this.rdates.remove(value as DateInput<T>);
+        rdates = this.rdates.remove(value as DateInput<T>) as RD;
         break;
       case 'exdate':
         exdates = this.exdates.remove(value as DateInput<T>);
@@ -245,10 +267,10 @@ export class VEvent<T extends typeof DateAdapter, D = any> extends OccurrenceGen
     });
   }
 
-  set(prop: 'timezone', value: string | null): VEvent<T, D>;
-  set(prop: 'start', value: DateInput<T>): VEvent<T, D>;
-  set(prop: 'rrules' | 'exrules', value: Rule<T, unknown>[]): VEvent<T, D>;
-  set(prop: 'rdates' | 'exdates', value: Dates<T, unknown>): VEvent<T, D>;
+  set(prop: 'timezone', value: string | null): VEvent<T, D, RR, RD>;
+  set(prop: 'start', value: DateInput<T>): VEvent<T, D, RR, RD>;
+  set(prop: 'rrules' | 'exrules', value: Rule<T, unknown>[]): VEvent<T, D, RR, RD>;
+  set(prop: 'rdates' | 'exdates', value: Dates<T, unknown>): VEvent<T, D, RR, RD>;
   set(
     prop: 'start' | 'timezone' | 'rrules' | 'exrules' | 'rdates' | 'exdates',
     value: DateInput<T> | string | null | Rule<T, unknown>[] | Dates<T, unknown>,
@@ -276,13 +298,13 @@ export class VEvent<T extends typeof DateAdapter, D = any> extends OccurrenceGen
         break;
       }
       case 'rrules':
-        rrules = value as Rule<T, unknown>[];
+        rrules = value as RR[];
         break;
       case 'exrules':
         exrules = value as Rule<T, unknown>[];
         break;
       case 'rdates':
-        rdates = value as Dates<T, unknown>;
+        rdates = value as RD;
         break;
       case 'exdates':
         exdates = value as Dates<T, unknown>;
@@ -312,7 +334,7 @@ export class VEvent<T extends typeof DateAdapter, D = any> extends OccurrenceGen
     let index = 0;
 
     while (date && (count === undefined || count > index)) {
-      date.generators.push(this);
+      date.generators.unshift(this);
 
       const yieldArgs = yield this.normalizeRunOutput(date);