+++ packages/rschedule/src/lib/operators/occurrence-stream.ts
@@ -1,4 +1,3 @@
-import { ArgumentError } from '../basic-utilities';
 import { DateAdapter } from '../date-adapter';
 import { DateTime } from '../date-time';
 import { IRunArgs, OccurrenceGenerator } from '../interfaces';
@@ -58,7 +57,9 @@ export class OccurrenceStream<T extends typeof DateAdapter> extends OccurrenceGe
    * of determining if an object is an `OccurrenceStream` by checking against the
    * global symbol registry.
    */
-  static isOccurrenceStream(object: unknown): object is OccurrenceStream<any> {
+  static isOccurrenceStream<T extends typeof DateAdapter = any>(
+    object: unknown,
+  ): object is OccurrenceStream<T> {
     return !!(object && typeof object === 'object' && (object as any)[OCCURRENCE_STREAM_ID]);
   }