+++ packages/rschedule/src/lib/schedule/schedule.ts
@@ -1,28 +1,44 @@
+import { Include, TupleUnshift } from '../basic-utilities';
 import { DateAdapter } from '../date-adapter';
 import { DateTime } from '../date-time';
 import { Dates } from '../dates';
 import { IRunArgs, IScheduleLike, OccurrenceGenerator } from '../interfaces';
+import {
+  CollectionIterator,
+  ICollectionsArgs,
+  IOccurrencesArgs,
+  OccurrenceIterator,
+} from '../iterators';
 import { add, OccurrenceStream, OperatorFnOutput, pipeFn, subtract, unique } from '../operators';
 import { IProvidedRuleOptions, Rule } from '../rule';
 import { DateInput } from '../utilities';
 
 const SCHEDULE_ID = Symbol.for('35d5d3f8-8924-43d2-b100-48e04b0cf500');
 
-export class Schedule<T extends typeof DateAdapter, D = any> extends OccurrenceGenerator<T>
-  implements IScheduleLike<T> {
+type GetScheduleType<T> = Include<T, Schedule<any, any, any, any>> extends never
+  ? Schedule<any, any, any, any>
+  : Include<T, Schedule<any, any, any, any>>;
+
+export class Schedule<
+  T extends typeof DateAdapter,
+  D = any,
+  RR extends Rule<T> = Rule<T>,
+  RD extends Dates<T> = Dates<T>
+> extends OccurrenceGenerator<T> implements IScheduleLike<T> {
   /**
    * Similar to `Array.isArray`, `isSchedule` provides a surefire method
    * of determining if an object is a `Schedule` by checking against the
    * global symbol registry.
    */
-  static isSchedule(object: any): object is Schedule<any> {
+  // @ts-ignore the check is working as intended but typescript doesn't like it for some reason
+  static isSchedule<T>(object: T): object is GetScheduleType<T> {
     return !!(object && typeof object === 'object' && (object as any)[SCHEDULE_ID]);
   }
 
   // For some reason, error is thrown if typed as `readonly Rule<T>[]`
-  readonly rrules: ReadonlyArray<Rule<T>> = [];
+  readonly rrules: ReadonlyArray<RR> = [];
   readonly exrules: ReadonlyArray<Rule<T>> = [];
-  readonly rdates: Dates<T>;
+  readonly rdates: RD;
   readonly exdates: Dates<T>;
 
   pipe: (...operatorFns: OperatorFnOutput<T>[]) => OccurrenceStream<T> = pipeFn(this);
@@ -42,9 +58,9 @@ export class Schedule<T extends typeof DateAdapter, D = any> extends OccurrenceG
       dateAdapter?: T;
       timezone?: string | null;
       data?: D;
-      rrules?: ReadonlyArray<IProvidedRuleOptions<T> | Rule<T>>;
+      rrules?: ReadonlyArray<IProvidedRuleOptions<T> | RR>;
       exrules?: ReadonlyArray<IProvidedRuleOptions<T> | Rule<T>>;
-      rdates?: ReadonlyArray<DateInput<T>> | Dates<T>;
+      rdates?: ReadonlyArray<DateInput<T>> | RD;
       exdates?: ReadonlyArray<DateInput<T>> | Dates<T>;
     } = {},
   ) {
@@ -57,12 +73,12 @@ export class Schedule<T extends typeof DateAdapter, D = any> extends OccurrenceG
     if (args.rrules) {
       this.rrules = args.rrules.map(ruleArgs => {
         if (Rule.isRule(ruleArgs)) {
-          return ruleArgs.set('timezone', this.timezone);
+          return ruleArgs.set('timezone', this.timezone) as RR;
         } else {
           return new Rule(ruleArgs as IProvidedRuleOptions<T>, {
             dateAdapter: this.dateAdapter as any,
             timezone: this.timezone,
-          });
+          }) as RR;
         }
       });
     }
@@ -82,14 +98,17 @@ export class Schedule<T extends typeof DateAdapter, D = any> extends OccurrenceG
 
     if (args.rdates) {
       this.rdates = Dates.isDates(args.rdates)
-        ? args.rdates.set('timezone', this.timezone)
-        : new Dates({
+        ? (args.rdates.set('timezone', this.timezone) as RD)
+        : (new Dates({
             dates: args.rdates as ReadonlyArray<DateInput<T>>,
             dateAdapter: this.dateAdapter as any,
             timezone: this.timezone,
-          });
+          }) as RD);
     } else {
-      this.rdates = new Dates({ dateAdapter: this.dateAdapter as any, timezone: this.timezone });
+      this.rdates = new Dates({
+        dateAdapter: this.dateAdapter as any,
+        timezone: this.timezone,
+      }) as RD;
     }
 
     if (args.exdates) {
@@ -125,8 +144,16 @@ export class Schedule<T extends typeof DateAdapter, D = any> extends OccurrenceG
     });
   }
 
-  add(prop: 'rrule' | 'exrule', value: Rule<T, unknown>): Schedule<T, D>;
-  add(prop: 'rdate' | 'exdate', value: DateInput<T>): Schedule<T, D>;
+  occurrences(args: IOccurrencesArgs<T> = {}): OccurrenceIterator<T, [this, RR | RD]> {
+    return new OccurrenceIterator(this, this.normalizeOccurrencesArgs(args));
+  }
+
+  collections(args: ICollectionsArgs<T> = {}): CollectionIterator<T, [this, RR | RD]> {
+    return new CollectionIterator(this, this.normalizeCollectionsArgs(args));
+  }
+
+  add(prop: 'rrule' | 'exrule', value: Rule<T, unknown>): Schedule<T, D, RR, RD>;
+  add(prop: 'rdate' | 'exdate', value: DateInput<T>): Schedule<T, D, RR, RD>;
   add(prop: 'rdate' | 'exdate' | 'rrule' | 'exrule', value: Rule<T, unknown> | DateInput<T>) {
     const rrules = this.rrules.slice();
     const exrules = this.exrules.slice();
@@ -135,13 +162,13 @@ export class Schedule<T extends typeof DateAdapter, D = any> extends OccurrenceG
 
     switch (prop) {
       case 'rrule':
-        rrules.push(value as Rule<T, unknown>);
+        rrules.push(value as RR);
         break;
       case 'exrule':
         exrules.push(value as Rule<T, unknown>);
         break;
       case 'rdate':
-        rdates = this.rdates.add(value as DateInput<T>);
+        rdates = this.rdates.add(value as DateInput<T>) as RD;
         break;
       case 'exdate':
         exdates = this.exdates.add(value as DateInput<T>);
@@ -159,8 +186,8 @@ export class Schedule<T extends typeof DateAdapter, D = any> extends OccurrenceG
     });
   }
 
-  remove(prop: 'rrule' | 'exrule', value: Rule<T, unknown>): Schedule<T, D>;
-  remove(prop: 'rdate' | 'exdate', value: DateInput<T>): Schedule<T, D>;
+  remove(prop: 'rrule' | 'exrule', value: Rule<T, unknown>): Schedule<T, D, RR, RD>;
+  remove(prop: 'rdate' | 'exdate', value: DateInput<T>): Schedule<T, D, RR, RD>;
   remove(prop: 'rdate' | 'exdate' | 'rrule' | 'exrule', value: Rule<T, unknown> | DateInput<T>) {
     let rrules = this.rrules;
     let exrules = this.exrules;
@@ -175,7 +202,7 @@ export class Schedule<T extends typeof DateAdapter, D = any> extends OccurrenceG
         exrules = exrules.filter(rule => rule !== value);
         break;
       case 'rdate':
-        rdates = this.rdates.remove(value as DateInput<T>);
+        rdates = this.rdates.remove(value as DateInput<T>) as RD;
         break;
       case 'exdate':
         exdates = this.exdates.remove(value as DateInput<T>);
@@ -193,9 +220,9 @@ export class Schedule<T extends typeof DateAdapter, D = any> extends OccurrenceG
     });
   }
 
-  set(prop: 'timezone', value: string | null): Schedule<T, D>;
-  set(prop: 'rrules' | 'exrules', value: Rule<T, unknown>[]): Schedule<T, D>;
-  set(prop: 'rdates' | 'exdates', value: Dates<T, unknown>): Schedule<T, D>;
+  set(prop: 'timezone', value: string | null): Schedule<T, D, RR, RD>;
+  set(prop: 'rrules' | 'exrules', value: Rule<T, unknown>[]): Schedule<T, D, RR, RD>;
+  set(prop: 'rdates' | 'exdates', value: Dates<T, unknown>): Schedule<T, D, RR, RD>;
   set(
     prop: 'timezone' | 'rrules' | 'exrules' | 'rdates' | 'exdates',
     value: string | null | Rule<T, unknown>[] | Dates<T, unknown>,
@@ -212,13 +239,13 @@ export class Schedule<T extends typeof DateAdapter, D = any> extends OccurrenceG
         timezone = value as string | null;
         break;
       case 'rrules':
-        rrules = value as Rule<T, unknown>[];
+        rrules = value as RR[];
         break;
       case 'exrules':
         exrules = value as Rule<T, unknown>[];
         break;
       case 'rdates':
-        rdates = value as Dates<T, unknown>;
+        rdates = value as RD;
         break;
       case 'exdates':
         exdates = value as Dates<T, unknown>;
@@ -248,7 +275,7 @@ export class Schedule<T extends typeof DateAdapter, D = any> extends OccurrenceG
     let index = 0;
 
     while (date && (count === undefined || count > index)) {
-      date.generators.push(this);
+      date.generators.unshift(this);
 
       const yieldArgs = yield this.normalizeRunOutput(date);